requirements.txt
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-tavily>=0.2.0
langchain-community>=0.3.0
google-generativeai
requests
python-dotenv
